optimizer/requirements.txt
# JJC Bot Optimizer Dependencies

# Core dependencies
python-dotenv>=1.0.0

# LLM providers (install based on your choice)
anthropic>=0.18.0      # For Claude API
openai>=1.12.0         # For OpenAI API
requests>=2.31.0       # For Ollama (local models)

# Optional but recommended
# MetaTrader5>=5.0.45  # MT5 Python package (Windows only, optional)
